feat: Add photo attachments to environment log entries
This commit introduces the ability to attach photos to environment log entries. This includes database schema updates, UI changes in the environment log screen, and updates to the repository to handle photo persistence and cleanup.

A new database migration (`15_16`) is added to include a `photoUri` text column to the `EnvironmentEntry` table.

- **`PlantRepository.java`**:
    - Manages the lifecycle of environment photos. It now saves a copy of the selected photo to a dedicated app-private directory (`environment_photos`) upon insertion or update of an `EnvironmentEntry`.
    - When an entry is updated with a new photo, the old photo file is deleted.
    - When an entry is deleted, its associated photo is also deleted.
    - `updateEnvironmentEntry` and `deleteEnvironmentEntry` now accept an optional `previousPhotoUri` to correctly handle photo cleanup.

- **`feature/environment`**:
    - `EnvironmentLogFragment.java`: The fragment now handles photo selection via `ActivityResultLauncher` and camera results. It communicates photo selections and removals to the presenter.
    - `EnvironmentLogPresenter.java`: Manages the state of a pending photo (`pendingPhotoUri`) for new or edited entries. It passes the photo URI to the repository for persistence. An entry can now be saved if it has a photo, even if all other fields are empty.
    - `EnvironmentLogView.java`: A new `showPhotoPreview` method is added to the view interface to update the UI with the selected photo.

- **`data/util/PhotoManager.java`**:
    - A new `saveEnvironmentPhoto` method is added to store photos in a separate `environment_photos` directory.
    - `isEnvironmentPhoto` helps distinguish these photos from other app-managed images.

- **Database & Migration**:
    - `PlantDatabase.java`:
        - The database version is bumped to 16.
        - `MIGRATION_15_16` is added to alter the `EnvironmentEntry` table and add the `photoUri` column.
    - `PlantDatabaseMigrationTest.java`:
        - Added `migrate15To16_addsEnvironmentPhotoColumn` to verify the new migration correctly adds the `photoUri` column and sets its value to `NULL` for existing entries.

- **Testing**:
    - `EnvironmentLogInstrumentedTest.java`:
        - New instrumented test `attachPhotoShowsPreviewAndSaves` verifies the end-to-end flow of attaching a photo from the gallery, seeing the preview, saving the entry, and confirming the photo is displayed in the log.
        - `espresso-intents` library was added for mocking the gallery intent.
    - `PlantRepositoryTest.java`:
        - New test `environmentEntryPhotoLifecycle` verifies that photos are correctly saved on insert, replaced on update (deleting the old one), and deleted when an entry is removed.
    - `EnvironmentLogPresenterTest.java`:
        - Added `onSubmit_withOnlyPhotoPersistsEntry` to ensure an entry with only a photo can be successfully saved.

+++ app/src/test/java/de/oabidi/pflanzenbestandundlichttest/CareRecommendationEngineTest.java
@@ -124,7 +124,7 @@ public class CareRecommendationEngineTest {
             emptyStage,
             new SpeciesTarget.WateringInfo(null, null, null),
             new SpeciesTarget.FloatRange(temperatureMin, temperatureMax),
-            new SpeciesTarget.FloatRange((Float) 40.0, (Float) 60.0),
+            new SpeciesTarget.FloatRange(40f, 60f),
             null,
             null,
             null,
